Add files via upload
GlobePics Journey is an innovative web application that merges the art of photography with the science of geospatial technology. Utilizing interactive mapping tools like Leaflet.js and Mapbox, it offers users a platform to share and discover photos linked to specific geographic locations. Backed by PostgreSQL/PostGIS for robust data management and featuring a sleek interface built with HTML, CSS, and JavaScript, it stands as a testament to the power of Web GIS in bridging distances and bringing global communities together. This repository contains the complete codebase and deployment configuration for the GlobePics Journey application.